added more operators
previously, i thought it might be ambiguous if ++/--/+/-/+=/-= would be ambiguous as to if they affect the pointer or the meta data, so they were excluded.
however, now i think it's obvious given that 'ptr' is in the name of the class, so a full set of pointer arithmetic operators has been added.